refactor(vmpool): gate VirtualMachinePool at runtime, not via a build tag
Paid-edition features in this module (VolumeMigration, USB, the hotplug ones)
are ordinary Apache-licensed code compiled in every edition and gated at runtime
by their feature gate; only tiny edition shims use //go:build EE. VirtualMachinePool
was the odd one out — its controller and apiserver code was //go:build EE, which
kept it out of the default (CE) unit-test run and forced the whole suite to build
with -tags EE, in turn breaking edition-default tests elsewhere (e.g. vmop's
locked-feature case).

Align it with the rest: drop the EE build tag (Apache headers), rely on the
existing VirtualMachinePool feature gate (locked off in CE) that SetupController
and the scale webhook already check, collapse the ee/ce setup and apiserver-install
shims into single files, and revert the test:unit and virtualization-api build
tweaks. No behaviour change in EE; the code is simply inert in CE.
