+++ images/virtualization-artifact/Taskfile.yaml
@@ -52,9 +52,7 @@ tasks:
     desc: "Run go unit tests"
     cmds:
       - |
-        # Build with the EE tag: EE is the default shipped edition (see werf.inc.yaml,
-        # MODULE_EDITION defaults to EE), so unit tests must cover //go:build EE code too.
-        go tool ginkgo -v --tags EE -r pkg/
+        go tool ginkgo -v -r pkg/
 
   lint:
     desc: "Run linters locally"
