--- images/virtualization-artifact/pkg/apiserver/api/install_enterprise_ce.go
@@ -1,29 +1,0 @@
-//go:build !EE
-// +build !EE
-
-/*
-Copyright 2026 Flant JSC
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-     http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package api
-
-import (
-	"k8s.io/apiserver/pkg/registry/rest"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-)
-
-// installEnterpriseResources is a no-op in CE builds: paid-edition subresources
-// (e.g. virtualmachinepools/scaledownwith) are not compiled here.
-func installEnterpriseResources(_ map[string]rest.Storage, _ client.Client) {}
